requirements.txt
# Core MCP and Goose integration
mcp>=1.0.0
uvicorn>=0.24.0
fastapi>=0.104.0

# Square API integration
squareup>=30.0.0

# Voice integration (ElevenLabs)
elevenlabs>=0.2.26

# Gesture detection (MediaPipe)
mediapipe>=0.10.7
opencv-python>=4.8.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
structlog>=23.2.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
